Put History class in separate module.
