+++ phy/cluster/manual/clustering.py
@@ -12,108 +12,7 @@ import numpy as np
 
 from ...ext.six import iterkeys, itervalues
 from ._utils import _unique, _spikes_in_clusters
-
-
-#------------------------------------------------------------------------------
-# History class
-#------------------------------------------------------------------------------
-
-class History(object):
-    """Implement a history of actions with an undo stack."""
-    def __init__(self, base_item=None):
-        self.clear(base_item)
-
-    def clear(self, base_item=None):
-        """Clear the history."""
-        # List of changes, contains at least the base item.
-        self._history = [base_item]
-        # Index of the current item.
-        self._index = 0
-
-    @property
-    def current_item(self):
-        """Return the current element."""
-        if self._history and self._index >= 0:
-            self._check_index()
-            return self._history[self._index]
-
-    @property
-    def current_position(self):
-        """Current position in the history."""
-        return self._index
-
-    def _check_index(self):
-        """Check that the index is without the bounds of _history."""
-        assert 0 <= self._index <= len(self._history) - 1
-        # There should always be the base item at least.
-        assert len(self._history) >= 1
-
-    def iter(self, start=0, end=None):
-        """Iterate through successive history items.
-
-        Parameters
-        ----------
-        end : int
-            Index of the last item to loop through + 1.
-
-        start : int
-            Initial index for the loop (0 by default).
-
-        """
-        if end is None:
-            end = self._index + 1
-        elif end == 0:
-            raise StopIteration()
-        if start >= end:
-            raise StopIteration()
-        # Check arguments.
-        assert 0 <= end <= len(self._history)
-        assert 0 <= start <= end - 1
-        for i in range(start, end):
-            yield self._history[i]
-
-    def __iter__(self):
-        return self.iter()
-
-    def __len__(self):
-        return len(self._history)
-
-    def add(self, item):
-        """Add an item in the history."""
-        self._check_index()
-        # Possibly truncate the history up to the current point.
-        self._history = self._history[:self._index + 1]
-        # Append the item
-        self._history.append(item)
-        # Increment the index.
-        self._index += 1
-        self._check_index()
-        # Check that the current element is what was provided to the function.
-        assert id(self.current_item) == id(item)
-
-    def back(self):
-        """Go back in history if possible.
-
-        Return the current item after going back.
-
-        """
-        if self._index <= 0:
-            return None
-        self._index -= 1
-        self._check_index()
-        return self.current_item
-
-    def forward(self):
-        """Go forward in history if possible.
-
-        Return the current item after going forward.
-
-        """
-        if self._index >= len(self._history) - 1:
-            return None
-        self._index += 1
-        self._check_index()
-        return self.current_item
+from ._history import History
 
 
 #------------------------------------------------------------------------------
@@ -238,10 +137,10 @@ class Clustering(object):
 
 
 #------------------------------------------------------------------------------
-# ClusterManager class
+# ClusterMetadata class
 #------------------------------------------------------------------------------
 
-class ClusterManager(object):
+class ClusterMetadata(object):
     """Object holding cluster metadata.
 
     Constructor
